[llubi] Move llubi specific source/header files into ./lib
